backend/requirements.txt
Flask==3.1.0
flask-cors==5.0.1
google-cloud-firestore==2.20.1
gunicorn==23.0.0
python-dotenv==1.0.1
